Move Images struct to pkg/api/pipeline 🏃
This allows to use that struct from other package more safely, aka
without depending on reconciler package(s).
