+++ pkg/reconciler/reconciler.go
@@ -19,6 +19,7 @@ package reconciler
 import (
 	"time"
 
+	"github.com/tektoncd/pipeline/pkg/apis/pipeline"
 	clientset "github.com/tektoncd/pipeline/pkg/client/clientset/versioned"
 	pipelineScheme "github.com/tektoncd/pipeline/pkg/client/clientset/versioned/scheme"
 	"go.uber.org/zap"
@@ -81,12 +82,12 @@ type Base struct {
 	Logger *zap.SugaredLogger
 
 	// Images contains images to use for certain internal container
-	Images Images
+	Images pipeline.Images
 }
 
 // NewBase instantiates a new instance of Base implementing
 // the common & boilerplate code between our reconcilers.
-func NewBase(opt Options, controllerAgentName string, images Images) *Base {
+func NewBase(opt Options, controllerAgentName string, images pipeline.Images) *Base {
 	// Enrich the logs with controller name
 	logger := opt.Logger.Named(controllerAgentName).With(zap.String(logkey.ControllerType, controllerAgentName))
